# Use unsafe_get/put_pixel