--- a/src/map.rs
+++ b/src/map.rs
@@ -57,7 +57,8 @@ pub fn map_subpixels<I, P, F, S>(image: &I, f: F) -> VecBuffer<ChannelMap<P, S>>
             let mut out_channels = out.get_pixel_mut(x, y).channels_mut();
             for c in 0..P::channel_count() {
                 out_channels[c as usize]
-                    = f(image.get_pixel(x, y).channels()[c as usize]);
+                    = f(unsafe {*image.unsafe_get_pixel(x, y)
+                        .channels().get_unchecked(c as usize) });
             }
         }
     }
@@ -77,8 +78,10 @@ pub fn map_colors<I, P, Q, F>(image: &I, f: F) -> VecBuffer<Q>
 
     for y in 0..height {
         for x in 0..width {
-            let pix = image.get_pixel(x, y);
-            out.put_pixel(x, y, f(pix));
+            unsafe {
+                let pix = image.unsafe_get_pixel(x, y);
+                out.unsafe_put_pixel(x, y, f(pix));
+            }
         }
     }
 
@@ -97,8 +100,10 @@ pub fn map_pixels<I, P, Q, F>(image: &I, f: F) -> VecBuffer<Q>
 
     for y in 0..height {
         for x in 0..width {
-            let pix = image.get_pixel(x, y);
-            out.put_pixel(x, y, f(x, y, pix));
+            unsafe {
+                let pix = image.unsafe_get_pixel(x, y);
+                out.unsafe_put_pixel(x, y, f(x, y, pix));
+            }
         }
     }