This PR change most `get_pixel`/`put_pixel` into their unsafe equivalent `unsafe_get_pixel`/`unsafe_put_pixel`.

These functions are unsafe because there is no bound checks. However those bounds check are most of the time already enforced within the fns by looping in 0..height x 0..width.

It results in a very nice speedup in most benches but adds a huge amount of `unsafe`s. We may probably need to refactor in a the future so only very few functions do use unsafes.

There is also a simplification of the code for `suppress_non_maximum_mut` in a separate commit.

```
before
test affine::test::bench_rotate_bilinear                              ... bench:     665,256 ns/iter (+/- 5,648)
test affine::test::bench_rotate_nearest                               ... bench:     356,280 ns/iter (+/- 2,612)
test affine::test::bench_translate                                    ... bench:     507,632 ns/iter (+/- 4,090)
test contrast::test::bench_equalize_histogram                         ... bench:   3,526,780 ns/iter (+/- 20,114)
test corners::test::bench_is_corner_fast12_12_noncontiguous           ... bench:          27 ns/iter (+/- 0)
test corners::test::bench_is_corner_fast9_9_contiguous_lighter_pixels ... bench:          31 ns/iter (+/- 0)
test filter::test::bench_box_filter                                   ... bench:   4,587,544 ns/iter (+/- 20,103)
test filter::test::bench_filter3x3_i32_filter                         ... bench:   2,557,473 ns/iter (+/- 13,557)
test filter::test::bench_horizontal_filter                            ... bench:   9,357,790 ns/iter (+/- 141,952)
test filter::test::bench_separable_filter                             ... bench:   6,631,563 ns/iter (+/- 39,749)
test filter::test::bench_vertical_filter                              ... bench:   9,385,985 ns/iter (+/- 39,437)
test integralimage::test::bench_column_running_sum                    ... bench:       3,761 ns/iter (+/- 43)
test integralimage::test::bench_integral_image                        ... bench:   1,446,309 ns/iter (+/- 7,460)
test integralimage::test::bench_row_running_sum                       ... bench:       3,432 ns/iter (+/- 7)
test suppress::test::bench_local_maxima_dense                         ... bench:      45,662 ns/iter (+/- 947)
test suppress::test::bench_local_maxima_sparse                        ... bench:      44,744 ns/iter (+/- 206)
test suppress::test::bench_suppress_non_maximum_decreasing_gradient   ... bench:     182,820 ns/iter (+/- 1,010)
test suppress::test::bench_suppress_non_maximum_increasing_gradient   ... bench:     200,905 ns/iter (+/- 3,371)
test suppress::test::bench_suppress_non_maximum_noise                 ... bench:     228,426 ns/iter (+/- 2,115)


after
test affine::test::bench_rotate_bilinear                              ... bench:     631,805 ns/iter (+/- 8,626)
test affine::test::bench_rotate_nearest                               ... bench:     349,432 ns/iter (+/- 3,367)
test affine::test::bench_translate                                    ... bench:     224,115 ns/iter (+/- 1,665)
test contrast::test::bench_equalize_histogram                         ... bench:   2,350,830 ns/iter (+/- 17,506)
test corners::test::bench_is_corner_fast12_12_noncontiguous           ... bench:          24 ns/iter (+/- 0)
test corners::test::bench_is_corner_fast9_9_contiguous_lighter_pixels ... bench:          25 ns/iter (+/- 0)
test filter::test::bench_box_filter                                   ... bench:   4,349,047 ns/iter (+/- 9,687)
test filter::test::bench_filter3x3_i32_filter                         ... bench:   1,941,680 ns/iter (+/- 8,936)
test filter::test::bench_horizontal_filter                            ... bench:   9,162,881 ns/iter (+/- 32,956)
test filter::test::bench_separable_filter                             ... bench:   6,583,317 ns/iter (+/- 106,594)
test filter::test::bench_vertical_filter                              ... bench:   9,382,289 ns/iter (+/- 42,686)
test integralimage::test::bench_column_running_sum                    ... bench:       3,573 ns/iter (+/- 64)
test integralimage::test::bench_integral_image                        ... bench:   1,331,922 ns/iter (+/- 12,368)
test integralimage::test::bench_row_running_sum                       ... bench:       3,268 ns/iter (+/- 12)
test suppress::test::bench_local_maxima_dense                         ... bench:      46,297 ns/iter (+/- 7,117)
test suppress::test::bench_local_maxima_sparse                        ... bench:      44,729 ns/iter (+/- 150)
test suppress::test::bench_suppress_non_maximum_decreasing_gradient   ... bench:     116,466 ns/iter (+/- 478)
test suppress::test::bench_suppress_non_maximum_increasing_gradient   ... bench:     125,538 ns/iter (+/- 992)
test suppress::test::bench_suppress_non_maximum_noise                 ... bench:      78,835 ns/iter (+/- 390)
```