@tafia 2016-01-06
The rand crate doesn't compile.