@theotherphil 2016-01-06
Thanks! Having to use unsafe everywhere does feel like an admission that our abstractions aren't good enough, but as you say we can worry about that later, and the gains from your changes are huge. I'm planning on adding a more efficient version of suppress non-maximum soon (see paper linked to from the relevant issue), so it'd be a waste to put too much effort into the current version. I will be keeping the simpler version around in test code to compare to the new algorithm to, so the simplifications are still useful.