@theotherphil 2016-01-06
Apparently "Pull requests that have a failing status can’t be merged from a phone". The build's already broken by dependencies, so LGTM if you want to merge this yourself.