@tafia 2016-01-06
merged then!
For the unsafes I am looking into using more `ImageBuffer::from_fn` for instance. Not perfect yet.